Update cheshift.py
CheShift-2 server was updated, consequently the plug-in was changed to reflect those changes. Basically now the color code is different, and now the plug-in retrieves more files from the server than before. 3 pdb files (coloured differently) and one zip file containing detailed information of the protein validation and if available information to "repair flaws". All this changes will be explained in the PyMOL wiki ASAP.