proper search in multiple chains
Now the script works well when searching over multiple chains. Before the edit, the situation was the following. For example, we have a PDB file (for example, 3FAV) that contains multiple chains (A, B, C, etc.). We want to find some sequence (say, EAGN), but we do not know in which chain it will be found, or we are just too lazy to specify the chain. We run
findseq EAGN, 3fav
Imagine, the sequence was found in chain A, and the corresponding residues are 14-17. In this case, the selection returned will contain correct residues from chain A, but also residues 14-17 from chain B, chain C, etc. that differ in sequence completely.
With this edit, only the residues from the right chain are included into selection.